add gutenberg resource, refactoring
