+++ complexresource.go
@@ -1,0 +1,14 @@
+package main
+
+import "errors"
+
+func (website resource) searchComplexResource(query query) resource {
+	// check history
+
+	// query and analyze result
+	website.Error = errors.New("under construction")
+
+	// possibly update history
+
+	return website
+}
